RS/datamodel: Added serialization of directives.
