+++ server/prisma-rs/libs/datamodel/src/dml/validator/directive/builtin/default.rs
@@ -1,5 +1,5 @@
-use crate::dml;
 use crate::dml::validator::directive::{Args, DirectiveValidator, Error};
+use crate::{ast, dml};
 
 /// Prismas builtin `@default` directive.
 pub struct DefaultDirectiveValidator {}
@@ -21,5 +21,16 @@ impl DirectiveValidator<dml::Field> for DefaultDirectiveValidator {
         }
 
         return Ok(());
+    }
+
+    fn serialize(&self, field: &dml::Field) -> Result<Option<ast::Directive>, Error> {
+        if let Some(default_value) = &field.default_value {
+            return Ok(Some(ast::Directive::new(
+                self.directive_name(),
+                vec![ast::Argument::new("value", default_value.into())],
+            )));
+        }
+
+        Ok(None)
     }
 }
