+++ server/prisma-rs/libs/datamodel/src/dml/validator/directive/builtin/embedded.rs
@@ -1,5 +1,5 @@
-use crate::dml;
 use crate::dml::validator::directive::{Args, DirectiveValidator, Error};
+use crate::{ast, dml};
 
 /// Prismas builtin `@embedded` directive.
 pub struct EmbeddedDirectiveValidator {}
@@ -11,5 +11,13 @@ impl DirectiveValidator<dml::Model> for EmbeddedDirectiveValidator {
     fn validate_and_apply(&self, _args: &Args, obj: &mut dml::Model) -> Result<(), Error> {
         obj.is_embedded = true;
         return Ok(());
+    }
+
+    fn serialize(&self, model: &dml::Model) -> Result<Option<ast::Directive>, Error> {
+        if model.is_embedded {
+            return Ok(Some(ast::Directive::new(self.directive_name(), vec![])));
+        }
+
+        Ok(None)
     }
 }
