+++ server/prisma-rs/libs/datamodel/src/dml/validator/directive/builtin/relation.rs
@@ -1,6 +1,6 @@
 use crate::common::value::ValueListValidator;
-use crate::dml;
 use crate::dml::validator::directive::{Args, DirectiveValidator, Error};
+use crate::{ast, dml};
 
 /// Prismas builtin `@relation` directive.
 pub struct RelationDirectiveValidator {}
@@ -23,5 +23,27 @@ impl DirectiveValidator<dml::Field> for RelationDirectiveValidator {
         } else {
             return self.error("Invalid field type, not a relation.", &args.span());
         }
+    }
+
+    fn serialize(&self, field: &dml::Field) -> Result<Option<ast::Directive>, Error> {
+        if let dml::FieldType::Relation(relation_info) = &field.field_type {
+            let mut args = Vec::new();
+
+            if let Some(name) = &relation_info.name {
+                args.push(ast::Argument::new_string("name", &name));
+            }
+
+            let mut related_fields: Vec<ast::Value> = Vec::new();
+
+            for related_field in &relation_info.to_fields {
+                related_fields.push(ast::Value::StringValue(related_field.clone(), ast::Span::empty()));
+            }
+
+            args.push(ast::Argument::new_array("references", related_fields));
+
+            return Ok(Some(ast::Directive::new(self.directive_name(), args)));
+        }
+
+        Ok(None)
     }
 }
