+++ server/prisma-rs/libs/datamodel/src/dml/validator/directive/mod.rs
@@ -20,9 +20,13 @@ pub type Args<'a> = common::argument::Arguments<'a>;
 pub trait DirectiveValidator<T> {
     /// Gets the directive name.
     fn directive_name(&self) -> &str;
+
     /// Validates a directive and applies the directive
     /// to the given object.
     fn validate_and_apply(&self, args: &Args, obj: &mut T) -> Result<(), Error>;
+
+    /// Serilizes the given directive's arguments for rendering.
+    fn serialize(&self, obj: &T) -> Result<Option<ast::Directive>, Error>;
 
     /// Shorthand to construct an directive validation error.
     fn error(&self, msg: &str, span: &ast::Span) -> Result<(), Error> {
@@ -72,6 +76,9 @@ impl<T> DirectiveValidator<T> for DirectiveScope<T> {
     }
     fn validate_and_apply(&self, args: &Args, obj: &mut T) -> Result<(), Error> {
         self.inner.validate_and_apply(args, obj)
+    }
+    fn serialize(&self, obj: &T) -> Result<Option<ast::Directive>, Error> {
+        self.inner.serialize(obj)
     }
 }
